clippy fixes

+++ lib/vector-api-client/src/gql/metrics.rs
@@ -149,7 +149,7 @@ impl component_sent_events_throughputs_subscription::ComponentSentEventsThroughp
             .map(|output| {
                 (
                     output.output_id.clone(),
-                    output.throughput as i64,
+                    output.throughput,
                 )
             })
             .collect()
